+++ tsconfig.app.json
@@ -27,7 +27,8 @@
     /* 절대경로 설정 추가 */
     "baseUrl": "./src",
     "paths": {
-      "@/*": ["*"]
+      "@/*": ["*"],
+      "@styles/*": ["styles/*"]
     }
   },
   "include": ["src"]
