+++ vite.config.ts
@@ -14,6 +14,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': path.resolve(__dirname, 'src'),
+      '@styles': path.resolve(__dirname, './src/styles'),
     },
   },
   server: {
